+++ yf-core/src/cmd.h
@@ -52,7 +52,7 @@ typedef struct {
 typedef struct {
   YF_buffer buf;
   size_t offset;
-  short stride;
+  unsigned stride;
 } YF_cmd_ibuf;
 
 /* The parameters of a 'clear color' command. */
